refactor(auth): extract client IP logic to http util
move getClientIp method to common utils for reuse across modules
rename config to cfg for consistency
reorganize jwt strategy exports
