+++ prisma/seeds/products/index.ts
@@ -4,9 +4,9 @@
  * @LastEditors: 杨仕明 shiming.y@qq.com
  * @LastEditTime: 2026-01-13 14:58:47
  * @FilePath: /lulab_backend/prisma/seeds/products/index.ts
- * @Description: 
- * 
- * Copyright (c) 2026 by LuLab-Team, All Rights Reserved. 
+ * @Description:
+ *
+ * Copyright (c) 2026 by LuLab-Team, All Rights Reserved.
  */
 export { createProducts } from './products';
 export type { ProductConfig } from './type';
