+++ src/auth/auth.module.ts
@@ -2,7 +2,7 @@
  * @Author: 杨仕明 shiming.y@qq.com
  * @Date: 2025-10-01 06:58:19
  * @LastEditors: 杨仕明 shiming.y@qq.com
- * @LastEditTime: 2025-10-03 04:03:51
+ * @LastEditTime: 2026-01-14 10:21:11
  * @FilePath: /lulab_backend/src/auth/auth.module.ts
  * @Description:
  *
@@ -13,18 +13,14 @@ import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { ConfigModule } from '@nestjs/config';
-import type { ConfigType } from '@nestjs/config';
 import { AuthController } from './auth.controller';
 import { RegisterService } from './services/register.service';
 import { LoginService } from './services/login.service';
 import { PasswordService } from './services/password.service';
 import { TokenService } from './services/token.service';
 import { AuthPolicyService } from './services/auth-policy.service';
-import {
-  JwtStrategy,
-  JWT_USER_LOOKUP,
-  JWT_TOKEN_BLACKLIST,
-} from './strategies/jwt.strategy';
+import { JwtStrategy } from './strategies/jwt.strategy';
+import { JWT_USER_LOOKUP, JWT_TOKEN_BLACKLIST } from './types/jwt.types';
 import { RedisModule } from '@/redis/redis.module';
 import { MailModule } from '@/mail/mail.module';
 import { UserModule } from '@/user/user.module';
@@ -46,10 +42,10 @@ import { PermissionModule } from '@/permission/permission.module';
     ConfigModule.forFeature(jwtConfig),
     JwtModule.registerAsync({
       imports: [ConfigModule.forFeature(jwtConfig)],
-      useFactory: (config: ConfigType<typeof jwtConfig>) => ({
-        secret: config.accessSecret,
+      useFactory: (cfg: ReturnType<typeof jwtConfig>) => ({
+        secret: cfg.accessSecret,
         signOptions: {
-          expiresIn: config.accessExpiresIn,
+          expiresIn: cfg.accessExpiresIn,
         },
       }),
       inject: [jwtConfig.KEY],
