+++ src/auth/dto/refresh-token.dto.ts
@@ -2,12 +2,13 @@
  * @Author: 杨仕明 shiming.y@qq.com
  * @Date: 2025-12-28 11:37:14
  * @LastEditors: 杨仕明 shiming.y@qq.com
- * @LastEditTime: 2026-01-08 15:15:39
+ * @LastEditTime: 2026-01-14 11:26:55
  * @FilePath: /lulab_backend/src/auth/dto/refresh-token.dto.ts
  * @Description:
  *
  * Copyright (c) 2026 by LuLab-Team, All Rights Reserved.
  */
+
 import { ApiProperty } from '@nestjs/swagger';
 import { IsOptional, IsString, IsEnum } from 'class-validator';
 import { ClientType } from '@/auth/types/jwt.types';
